% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qtl.R
\name{breeding_values}
\alias{breeding_values}
\title{Adapted function from QTLpoly}
\usage{
breeding_values(qtl_info, probs, selected_mks, blups, beta.hat, pos)
}
\description{
Adapted function from QTLpoly
}
