% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{check_viewqtl_diaqtl_values}
\alias{check_viewqtl_diaqtl_values}
\title{Check viewqtl object for diaQTL uploaded files}
\usage{
check_viewqtl_diaqtl_values(viewqtl_obj, pos, ll, DIC, effect)
}
\arguments{
\item{viewqtl_obj}{an object of class \code{viewqtl}}

\item{pos}{sum of the position values to be matched}

\item{ll}{sum of the log likelihood values to be matched}

\item{DIC}{sum of the deltaDIC values to be matched}

\item{effect}{sum of the effects values to be matched}
}
\value{
if consistent, returns 0. If not consistent, returns a 
        vector with a number of tests, where \code{TRUE} indicates
        a failed test.
}
\description{
Check viewqtl object for diaQTL uploaded files
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
\keyword{internal}
