% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qtl.R
\name{coord_radar}
\alias{coord_radar}
\title{Change ggplot coordinates to plot radar - From package see}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{ariable to map angle to (x or y)}

\item{start}{offset of starting point from 12 o'clock in radians. Offset is applied clockwise or anticlockwise depending on value of direction.}

\item{direction}{1, clockwise; -1, anticlockwise}
}
\description{
Change ggplot coordinates to plot radar - From package see
}
\keyword{internal}
