% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{filter_non_conforming_classes}
\alias{filter_non_conforming_classes}
\title{Filter non-conforming classes in F1, non double reduced population.
Function from MAPpoly.}
\usage{
filter_non_conforming_classes(input.data, prob.thres = NULL)
}
\arguments{
\item{input.data}{object of class mappoly}

\item{prob.thres}{threshold for filtering genotypes by genotype probability values. If NULL, the filter is not applied.}
}
\value{
filtered \code{mappoly.data} object
}
\description{
Filter non-conforming classes in F1, non double reduced population.
Function from MAPpoly.
}
\keyword{internal}
