% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_map.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Color pallet ggplot-like - Adapted from MAPpoly}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
Color pallet ggplot-like - Adapted from MAPpoly
}
\keyword{internal}
