% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{imf_h}
\alias{imf_h}
\title{Map functions - from MAPpoly}
\usage{
imf_h(r)
}
\arguments{
\item{r}{vector with recombination fraction values}
}
\value{
vector with genetic distances
}
\description{
Map functions - from MAPpoly
}
\keyword{internal}
