% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{is.prob.data}
\alias{is.prob.data}
\title{Is it a probability dataset? Function from MAPpoly.}
\usage{
is.prob.data(x)
}
\arguments{
\item{x}{object of class \code{mappoly.data}}
}
\value{
TRUE/FALSE indicating if genotype probability information is present
}
\description{
Is it a probability dataset? Function from MAPpoly.
}
\keyword{internal}
