% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_map.R
\name{map_summary}
\alias{map_summary}
\title{Gets summary information from map.
Adapted from MAPpoly}
\usage{
map_summary(left.lim = 0, right.lim = 5, ch = 1, maps, d.p1, d.p2)
}
\arguments{
\item{left.lim}{covered window in the linkage map start position}

\item{ch}{linkage group ID}

\item{maps}{list containing a vector for each linkage group markers with marker positions (named with marker names)}

\item{d.p1}{list containing a data.frame for each group with parent 1 dosages. The data.frame contain the columns: 
1) character vector with chromosomes ID; 
2) Character vector with markers ID; 3) Character vector with parent ID; 
4) numerical vector with dosage}

\item{d.p2}{list containing a data.frame for each group with parent 2 dosages. See d.p1 parameter description}

\item{rigth.lim}{covered window in the linkage map end position}
}
\value{
list with linkage map information: doses;  number snps by group; cM per snp; map size; number of linkage groups
}
\description{
Gets summary information from map.
Adapted from MAPpoly
}
\keyword{internal}
