% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{mf_h}
\alias{mf_h}
\title{Haldane map function. From MAPpoly}
\usage{
mf_h(d)
}
\arguments{
\item{d}{vector containing recombination fraction values}
}
\value{
vector with genetic distances estimated with Haldane function
}
\description{
Haldane map function. From MAPpoly
}
\keyword{internal}
