% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{mrk_chisq_test}
\alias{mrk_chisq_test}
\title{Chi-square test. Function from MAPpoly.}
\usage{
mrk_chisq_test(x, ploidy)
}
\arguments{
\item{x}{data.frame containing dosage information}

\item{ploidy}{integer defining the specie ploidy}
}
\value{
vector with p-values for each marker
}
\description{
Chi-square test. Function from MAPpoly.
}
\keyword{internal}
