% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qtl.R
\name{only_plot_profile}
\alias{only_plot_profile}
\title{Only the plot part of plot_profile function}
\usage{
only_plot_profile(pl.in)
}
\arguments{
\item{pl.in}{output object from \code{plot_profile} when plot == TRUE}
}
\value{
ggplot graphic with QTL significance profile
}
\description{
Only the plot part of plot_profile function
}
\keyword{internal}
