% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_map.R
\name{plot_cm_mb}
\alias{plot_cm_mb}
\title{Scatter plot relating linkage map and genomic positions}
\usage{
plot_cm_mb(viewmap, group, range.min, range.max)
}
\arguments{
\item{viewmap}{object of class \code{viewmap}}

\item{group}{selected group ID}

\item{range.min}{minimum value of the selected position range}

\item{range.max}{maximum value of the selected position range}
}
\description{
Scatter plot relating linkage map and genomic positions
}
\keyword{internal}
