% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qtl.R
\name{plot_effects}
\alias{plot_effects}
\title{Plot effects data}
\usage{
plot_effects(
  data_effects.obj,
  software,
  design = c("bar", "circle", "digenic")
)
}
\arguments{
\item{data_effects.obj}{output of function \code{data_effects}}

\item{software}{character defining which software was used for QTL analysis. Currently support for: QTLpoly, diaQTL and polyqtlR.}

\item{design}{character defining the graphic design. Options: `bar` - barplot of the effects; 
`circle` - circular plot of the effects (useful to compare effects of different traits); 
`digenic` - heatmap plotting sum of additive effects (bottom diagonal) and digenic effects (top diagonal) when present}
}
\description{
Plot effects data
}
\keyword{internal}
