% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_map.R
\name{plot_one_map}
\alias{plot_one_map}
\title{Plot a single linkage group with no phase - from MAPpoly}
\usage{
plot_one_map(x, i = 0, horiz = FALSE, col = "lightgray")
}
\arguments{
\item{x}{vector of genetic distances}

\item{i}{margins size}

\item{horiz}{logical TRUE/FALSE. If TRUE the map is plotted horizontally.}

\item{col}{color pallete to be used}
}
\description{
Plot a single linkage group with no phase - from MAPpoly
}
\keyword{internal}
