% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{polyqtlReffects}
\alias{polyqtlReffects}
\title{Addapted from polyqtlR}
\usage{
polyqtlReffects(
  IBD_list,
  Phenotype.df,
  genotype.ID,
  trait.ID.c,
  linkage_group,
  LOD_data,
  cM_range = NULL
)
}
\description{
Addapted from polyqtlR
}
