% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_diaQTL}
\alias{prepare_diaQTL}
\title{Converts diaQTL output to viewqtl object}
\usage{
prepare_diaQTL(scan1_list, scan1_summaries_list, fitQTL_list, BayesCI_list)
}
\arguments{
\item{scan1_list}{list with results from diaQTL \code{scan1} function}

\item{scan1_summaries_list}{list with results from diaQTL \code{scan1_summaries} function}

\item{fitQTL_list}{list with results from diaQTL \code{fitQTL} function}

\item{BayesCI_list}{list with results from diaQTL \code{BayesCI} function}
}
\value{
object of class \code{viewqtl}
}
\description{
Converts diaQTL output to viewqtl object
}
\keyword{internal}
