% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_hidecan_examples}
\alias{prepare_hidecan_examples}
\title{Upload hidecan example files}
\usage{
prepare_hidecan_examples(example)
}
\arguments{
\item{example}{character indicating the example dataset selected}
}
\value{
object of class \code{viewpoly}
}
\description{
Upload hidecan example files
}
\keyword{internal}
