% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{prepare_map}
\alias{prepare_map}
\title{prepare maps for plot - from MAPpoly}
\usage{
prepare_map(input.map, config = "best")
}
\arguments{
\item{input.map}{object of class \code{mappoly.map}}

\item{config}{choose between 'best', 'all' or provide vector with defined configuration. 
'best' provide just the best estimated configuration. 'all' provides all possibles.}
}
\value{
list containing phase and dosage information
}
\description{
prepare maps for plot - from MAPpoly
}
\keyword{internal}
