% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_Mapdata}
\alias{prepare_Mapdata}
\title{Receive upload files and convert to viewmap object}
\usage{
prepare_Mapdata(
  mappoly_in = NULL,
  dosages = NULL,
  phases = NULL,
  genetic_map = NULL,
  example_map
)
}
\arguments{
\item{mappoly_in}{mappoly sequence object}

\item{dosages}{RDS compressed file with both parents dosage information.
It should contain four columns: 1) character vector with chromosomes ID; 
2) Character vector with markers ID; 3) Character vector with parent ID; 
4) numerical vector with dosage.}

\item{phases}{RDS compressed file with phases information. It should contain:
1) Character vector with chromosome ID; 2) Character vector with marker ID;
3 to (ploidy number)*2 columns with each parents haplotypes.}

\item{genetic_map}{RDS compressed file with the genetic map information}

\item{example_map}{if user don´t come in with any other input, use a example files.}
}
\description{
Receive upload files and convert to viewmap object
}
