% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_MAPpoly}
\alias{prepare_MAPpoly}
\title{Converts list of mappoly.map object into viewmap object}
\usage{
prepare_MAPpoly(mappoly_list)
}
\arguments{
\item{mappoly_list}{list with objects of class \code{mappoly.map}}
}
\value{
object of class \code{viewmap}
}
\description{
Converts list of mappoly.map object into viewmap object
}
\keyword{internal}
