% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_polymapR}
\alias{prepare_polymapR}
\title{Converts polymapR ouputs to viewmap object}
\usage{
prepare_polymapR(polymapR.dataset, polymapR.map, input.type, ploidy)
}
\arguments{
\item{polymapR.dataset}{a \code{polymapR} dataset}

\item{polymapR.map}{output map sequence from polymapR}

\item{input.type}{indicates whether the input is discrete ("disc") or probabilistic ("prob")}

\item{ploidy}{ploidy level}
}
\value{
object of class \code{viewmap}
}
\description{
Converts polymapR ouputs to viewmap object
}
\keyword{internal}
