% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_polyqtlR}
\alias{prepare_polyqtlR}
\title{Converts polyqtlR outputs to viewqtl object}
\usage{
prepare_polyqtlR(polyqtlR_QTLscan_list, polyqtlR_qtl_info, polyqtlR_effects)
}
\arguments{
\item{polyqtlR_QTLscan_list}{list containing results from polyqtlR \code{QTLscan_list} function}

\item{polyqtlR_qtl_info}{data.frame containing the QTL information:LG - group ID; Pos - QTL position (cM); 
pheno - phenotype ID; Pos_lower - lower position of confidence interval; Pos_upper - upper position of the confidence interval;
thresh - LOD threshold applied}

\item{polyqtlR_effects}{data.frame with results from \code{visualiseQTLeffects} polyqtlR function}
}
\value{
object of class \code{viewqtl}
}
\description{
Converts polyqtlR outputs to viewqtl object
}
\keyword{internal}
