% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_functions.R
\name{prepare_qtl}
\alias{prepare_qtl}
\title{take all information needed from qtlpoly objects}
\usage{
prepare_qtl(remim.mod, data)
}
\arguments{
\item{remim.mod}{object of class "qtlpoly.model" "qtlpoly.remim".}

\item{data}{object of class "qtlpoly.data"}
}
\description{
take all information needed from qtlpoly objects
}
