% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datas.R
\docType{data}
\name{qtl_bt}
\alias{qtl_bt}
\title{Sweet potato BT QTL data}
\format{
An object of class \code{"viewpoly.qtl"}.
}
\usage{
data(qtl_bt)
}
\description{
QTL results from an analysis made with random-effect model approach based on score statistics in 
a autohexaploid sweetpotato biparental population (‘Beauregard’ × ‘Tanzania’) 
with 315 full-sibs. This dataset only contains the genotype probabilities for five individuals 
of the population.
}
\references{
Pereira et al. (2020) Genetics, Volume 215, Issue 3
(\href{https://academic.oup.com/genetics/article/215/3/579/5930535}{Genetics})
}
\keyword{datasets}
