% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{read_custom_files}
\alias{read_custom_files}
\title{Read input custom format files}
\usage{
read_custom_files(dosages, phases, genetic_map)
}
\arguments{
\item{dosages}{RDS compressed file with both parents dosage information.
It should contain four columns: 1) character vector with chromosomes ID; 
2) Character vector with markers ID; 3) Character vector with parent ID; 
4) numerical vector with dosage.}

\item{phases}{RDS compressed file with phases information. It should contain:
1) Character vector with chromosome ID; 2) Character vector with marker ID;
3 to (ploidy number)*2 columns with each parents haplotypes.}

\item{genetic_map}{RDS compressed file with the genetic map information}
}
\description{
Read input custom format files
}
