% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{read_input_hidecan}
\alias{read_input_hidecan}
\title{Check hidecan inputs}
\usage{
read_input_hidecan(input_list, func)
}
\arguments{
\item{input_list}{shiny input result containing file path}

\item{func}{hidecan read input function}
}
\description{
Check hidecan inputs
}
