% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_map.R
\name{summary_maps}
\alias{summary_maps}
\title{Summary maps - adapted from MAPpoly}
\usage{
summary_maps(viewmap, software = NULL)
}
\arguments{
\item{viewmap}{a list of objects of class \code{viewmap}}

\item{software}{character defined from each software it comes from}
}
\value{
a data frame containing a brief summary of all maps
}
\description{
This function generates a brief summary table
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}

Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
\keyword{internal}
