% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datas.R
\docType{data}
\name{viewmap_hexa}
\alias{viewmap_hexa}
\title{Sweet potato BT map data}
\format{
An object of class \code{"viewpoly.map"}.
}
\usage{
data(qtl_bt)
}
\description{
Linakge map from an analysis made with autohexaploid sweetpotato biparental population (‘Beauregard’ × ‘Tanzania’) 
with 315 full-sibs.
}
\references{
Pereira et al. (2020) Genetics, Volume 215, Issue 3
(\href{https://academic.oup.com/genetics/article/215/3/579/5930535}{Genetics})
}
\keyword{datasets}
