% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datas.R
\docType{data}
\name{viewmap_tetra}
\alias{viewmap_tetra}
\title{Potato data of B2721 population - Atlantic x B1829-5}
\format{
An object of class \code{"viewpoly.map"}.
}
\usage{
data(viewmap_tetra)
}
\description{
The population comprises 156 offsprings genotyped with the SolCAP Infinium 8303 
potato array. The dataset also contains the genomic position of the SNPs from 
the Solanum tuberosum genome version 4.03.
}
\references{
Felcher KJ, Coombs JJ, Massa AN, Hansey CN, Hamilton JP, et al. (2012) 
Integration of Two Diploid Potato Linkage Maps with the Potato Genome Sequence. 
PLOS ONE 7(4): e36347. \href{https://doi.org/10.1371/journal.pone.0036347}{PLOS ONE}
}
\keyword{datasets}
