% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{imf_h}
\alias{imf_h}
\title{Map functions - from MAPpoly}
\usage{
imf_h(r)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Map functions - from MAPpoly
}
\keyword{internal}
