% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_examples}
\alias{prepare_examples}
\alias{prepare_map_custom_files}
\alias{prepare_MAPpoly}
\alias{prepare_polymapR}
\alias{prepare_QTLpoly}
\alias{prepare_diaQTL}
\alias{prepare_polyqtlR}
\alias{prepare_qtl_custom_files}
\title{Upload example files}
\usage{
prepare_examples(example)

prepare_map_custom_files(dosages, phases, genetic_map)

prepare_MAPpoly(mappoly_list)

prepare_polymapR(polymapR.dataset, polymapR.map, input.type, ploidy)

prepare_QTLpoly(data, remim.mod, est.effects, fitted.mod)

prepare_diaQTL(scan1_list, scan1_summaries_list, fitQTL_list, BayesCI_list)

prepare_polyqtlR(polyqtlR_QTLscan_list, polyqtlR_qtl_info, polyqtlR_effects)

prepare_qtl_custom_files(
  selected_mks,
  qtl_info,
  blups,
  beta.hat,
  profile,
  effects,
  probs
)
}
\arguments{
\item{example}{character indicating the example dataset selected}

\item{dosages}{TSV or TSV.GZ file with both parents dosage information.
It should contain four columns: 1) character vector with chromosomes ID; 
2) Character vector with markers ID; 3) Character vector with parent ID; 
4) numerical vector with dosage.}

\item{phases}{TSV or TSV.GZ file with phases information. It should contain:
1) Character vector with chromosome ID; 2) Character vector with marker ID;
3 to (ploidy number)*2 columns with each parents haplotypes.}

\item{genetic_map}{TSV or TSV.GZ file with the genetic map information}

\item{mappoly_list}{list with objects of class \code{mappoly.map}}

\item{polymapR.dataset}{a \code{polymapR} dataset}

\item{polymapR.map}{output map sequence from polymapR}

\item{input.type}{indicates whether the input is discrete ("disc") or probabilistic ("prob")}

\item{ploidy}{ploidy level}

\item{data}{object of class "qtlpoly.data"}

\item{remim.mod}{object of class "qtlpoly.model" "qtlpoly.remim".}

\item{est.effects}{object of class "qtlpoly.effects"}

\item{fitted.mod}{object of class "qtlpoly.fitted"}

\item{scan1_list}{list with results from diaQTL \code{scan1} function}

\item{scan1_summaries_list}{list with results from diaQTL \code{scan1_summaries} function}

\item{fitQTL_list}{list with results from diaQTL \code{fitQTL} function}

\item{BayesCI_list}{list with results from diaQTL \code{BayesCI} function}

\item{polyqtlR_QTLscan_list}{list containing results from polyqtlR \code{QTLscan_list} function}

\item{polyqtlR_qtl_info}{data.frame containing the QTL information:LG - group ID; Pos - QTL position (cM); 
pheno - phenotype ID; Pos_lower - lower position of confidence interval; Pos_upper - upper position of the confidence interval;
thresh - LOD threshold applied}

\item{polyqtlR_effects}{data.frame with results from \code{visualiseQTLeffects} polyqtlR function}

\item{selected_mks}{data.frame with: LG - linkage group ID; mk - marker ID; pos - position in linkage map (cM)}

\item{qtl_info}{data.frame with: LG - linkage group ID; Pos - position in linkage map (cM); 
Pheno - phenotype ID; Pos_lower - lower position of confidence interval; 
Pos_upper - upper position of the confidence interval; Pval - QTL p-value; h2 - herdability}

\item{blups}{data.frame with: haplo - haplotype ID; pheno - phenotype ID; qtl - QTL ID; u.hat - QTL estimated BLUPs}

\item{beta.hat}{data.frame with: pheno - phenotype ID; beta.hat - estimated beta}

\item{profile}{data.frame with: pheno - phenotype ID; LOP - significance value for the QTL, in this case LOP (can be LOD or DIC depending of the software used)}

\item{effects}{data.frame with: pheno - phenotype ID; qtl.id - QTL ID; haplo - haplotype ID; effect - haplotype effect value}

\item{probs}{data.frame with first column (named `ind`) as individuals ID and next columns named with markers ID and containing the genotype probability at each marker}
}
\value{
object of class \code{viewpoly}

object of class \code{viewmap}

object of class \code{viewmap}

object of class \code{viewmap}

object of class \code{viewqtl}

object of class \code{viewqtl}

object of class \code{viewqtl}

object of class \code{viewqtl}
}
\description{
Upload example files

Converts map information in custom format files to viewmap object

Converts list of mappoly.map object into viewmap object

Converts polymapR ouputs to viewmap object

Converts QTLpoly outputs to viewqtl object

Converts diaQTL output to viewqtl object

Converts polyqtlR outputs to viewqtl object

Converts QTL information in custom files to viewqtl object
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
\keyword{internal}
