% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{only_plot_profile}
\alias{only_plot_profile}
\title{Only the plot part of plot_profile function}
\usage{
only_plot_profile(
  lines,
  points,
  thre,
  sup.int,
  linesize,
  lgs.size,
  cutx,
  main,
  y.dat,
  by_range = FALSE
)
}
\description{
Only the plot part of plot_profile function
}
