% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qtl.R
\name{plot_effects}
\alias{plot_effects}
\title{Adapted function from QTLpoly}
\usage{
plot_effects(
  qtl_info,
  effects,
  pheno.col = NULL,
  p1 = "P1",
  p2 = "P2",
  df.info = NULL,
  lgs = NULL,
  groups = NULL,
  position = NULL,
  software
)
}
\description{
Adapted function from QTLpoly
}
