% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_diaQTL}
\alias{prepare_diaQTL}
\title{Prepare diaQTL output files}
\usage{
prepare_diaQTL(scan1_list, scan1_summaries_list, fitQTL_list, BayesCI_list)
}
\description{
Prepare diaQTL output files
}
