% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{prepare_map}
\alias{prepare_map}
\title{prepare maps for plot - from MAPpoly}
\usage{
prepare_map(input.map, config = "best")
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
prepare maps for plot - from MAPpoly
}
\keyword{internal}
