% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_MAPpoly}
\alias{prepare_MAPpoly}
\title{convert list of mappoly.map object into viewpoly_map object}
\usage{
prepare_MAPpoly(mappoly_list)
}
\description{
convert list of mappoly.map object into viewpoly_map object
}
