% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_polyqtlR}
\alias{prepare_polyqtlR}
\title{Upload polyqtlR outputs}
\usage{
prepare_polyqtlR(polyqtlR_QTLscan_list, polyqtlR_IBD, polyqtlR_phenotypes)
}
\description{
Upload polyqtlR outputs
}
