% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_QTLpoly}
\alias{prepare_QTLpoly}
\title{take all information needed from qtlpoly objects}
\usage{
prepare_QTLpoly(data, remim.mod, est.effects, fitted.mod)
}
\arguments{
\item{data}{object of class "qtlpoly.data"}

\item{remim.mod}{object of class "qtlpoly.model" "qtlpoly.remim".}
}
\description{
take all information needed from qtlpoly objects
}
