% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_upload.R
\name{prepare_QTLpoly}
\alias{prepare_QTLpoly}
\title{Converts QTLpoly outputs to viewqtl object}
\usage{
prepare_QTLpoly(data, remim.mod, est.effects, fitted.mod)
}
\arguments{
\item{data}{object of class "qtlpoly.data"}

\item{remim.mod}{object of class "qtlpoly.model" "qtlpoly.remim".}

\item{est.effects}{object of class "qtlpoly.effects"}

\item{fitted.mod}{object of class "qtlpoly.fitted"}
}
\value{
object of class \code{viewqtl}
}
\description{
Converts QTLpoly outputs to viewqtl object
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
\keyword{internal}
