% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qtl.R
\name{plot_profile}
\alias{plot_profile}
\alias{only_plot_profile}
\alias{data_effects}
\alias{coord_radar}
\alias{plot_effects}
\alias{breeding_values}
\alias{calc_homologprob}
\alias{plot.mappoly.homoprob}
\alias{select_haplo}
\title{Logarithm of \emph{P}-value (LOP) profile plots. Modified version of QTLpoly function.}
\usage{
plot_profile(
  profile,
  qtl_info,
  selected_mks,
  pheno.col = NULL,
  lgs.id = NULL,
  by_range = TRUE,
  range.min = NULL,
  range.max = NULL,
  plot = TRUE,
  software = NULL
)

only_plot_profile(pl.in)

data_effects(
  qtl_info,
  effects,
  pheno.col = NULL,
  p1 = "P1",
  p2 = "P2",
  lgs = NULL,
  groups = NULL,
  position = NULL,
  software,
  design = c("bar", "circle", "digenic")
)

coord_radar(theta = "x", start = 0, direction = 1)

plot_effects(
  data_effects.obj,
  software,
  design = c("bar", "circle", "digenic")
)

breeding_values(qtl_info, probs, selected_mks, blups, beta.hat, pos)

calc_homologprob(probs, selected_mks, selected_lgs)

\method{plot}{mappoly.homoprob}(x, stack = FALSE, lg = NULL, ind = NULL, verbose = TRUE, ...)

select_haplo(input.haplo, probs, selected_mks, effects.data)
}
\arguments{
\item{profile}{data.frame with: pheno - phenotype ID; LOP - significance value for the QTL. 
It can be LOP, LOD or DIC depending of the software used}

\item{qtl_info}{data.frame with: LG - linkage group ID; Pos - position in linkage map (cM); 
Pheno - phenotype ID; Pos_lower - lower position of confidence interval; 
Pos_upper - upper position of the confidence interval; Pval - QTL p-value; h2 - herdability}

\item{selected_mks}{data.frame with: LG - linkage group ID; mk - marker ID; pos - position in linkage map (cM)}

\item{pheno.col}{integer identifying phenotype}

\item{lgs.id}{integer identifying linkage group}

\item{by_range}{logical TRUE/FALSE. If TRUE range.min and range.max will set a colored window in the plot and the other positions will be gray.
If FALSE, range.min and range.max is ignored}

\item{range.min}{position in centimorgan defining the start of the colored window}

\item{range.max}{position in centimorgan defining the end of the colored window}

\item{plot}{logical TRUE/FALSE. If FALSE the function return a data.frame with information for \code{only_plot_profile} function. 
If TRUE, it returns a ggplot graphic.}

\item{software}{character defining which software was used for QTL analysis. Currently support for: QTLpoly, diaQTL and polyqtlR.}

\item{pl.in}{output object from \code{plot_profile} when plot == TRUE}

\item{effects}{data.frame with: pheno - phenotype ID; qtl.id - QTL ID; haplo - haplotype ID; effect - haplotype effect value}

\item{p1}{character with parent 1 ID}

\item{p2}{character with parent 2 ID}

\item{lgs}{vector of integers with linkage group ID of selected QTL/s}

\item{groups}{vector of integers with selected linkage group ID}

\item{position}{vector of centimorgan positions of selected QTL/s}

\item{design}{character defining the graphic design. Options: `bar` - barplot of the effects; 
`circle` - circular plot of the effects (useful to compare effects of different traits); 
`digenic` - heatmap plotting sum of additive effects (bottom diagonal) and digenic effects (top diagonal) when present}

\item{theta}{ariable to map angle to (x or y)}

\item{start}{offset of starting point from 12 o'clock in radians. Offset is applied clockwise or anticlockwise depending on value of direction.}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{data_effects.obj}{output of function \code{data_effects}}

\item{probs}{data.frame with first column (named `ind`) as individuals ID and next columns named with markers ID and containing the genotype probability at each marker}

\item{blups}{data.frame with: haplo - haplotype ID; pheno - phenotype ID; qtl - QTL ID; u.hat - QTL estimated BLUPs}

\item{beta.hat}{data.frame with: pheno - phenotype ID; beta.hat - estimated beta}

\item{pos}{selected QTL position (cM)}

\item{selected_lgs}{vector containing selected LGs ID}

\item{x}{an object of class \code{mappoly.homoprob}}

\item{stack}{logical. If \code{TRUE}, probability profiles of all homologues
are stacked in the plot (default = FALSE)}

\item{lg}{indicates which linkage group should be plotted. If \code{NULL} 
(default), it plots the first linkage group. If 
\code{"all"}, it plots all linkage groups}

\item{ind}{indicates which individuals should be plotted. It can be the 
position of the individuals in the dataset or it's name. 
If \code{NULL} (default), the function plots the first 
individual}

\item{verbose}{if \code{TRUE} (default), the current progress is shown; if
\code{FALSE}, no output is produced}

\item{...}{unused arguments}

\item{input.haplo}{character vector with selected haplotypes. It contains the information: "Trait:<trait ID>_LG:<linkage group ID_Pos:<QTL position>"}

\item{effects.data}{output object from \code{data_effects} function}
}
\value{
ggplot graphic (if plot == TRUE) or data.frame (if plot == FALSE) with information 
from QTL significance profile

ggplot graphic with QTL significance profile

ggplot graphic

data.frame containing breeding values

object of class \code{mappoly.homoprob}

ggplot graphic
}
\description{
Plots profiled logarithm of score-based \emph{P}-values (LOP) from individual or combined traits.
}
\keyword{internal}
